fixed carbon to float

+++ core/users/models.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import User
 
 class Profile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    emissions = models.IntegerField(default=0)
+    emissions = models.FloatField(default=0.0)
     notifications = models.BooleanField(default=False)
     notification_delay = models.IntegerField(default=5)
     age = models.TextField(default="adult")
